fix - ssr-error
1. ssr 적용
2. 게시판 - 테이블 분리
